move filelogger into own file
https://coveord.atlassian.net/browse/CDX-252
